+++ CHANGELOG.md
@@ -8,11 +8,15 @@ This changelog documents the changes applied to the previous version of this pro
 
 - this `CHANGELOG.md`
 - `requirements.txt` and information about their usage in `README.md`
+- docstrings for all functions
+- explicit typing for all functions
+- a pydantic Song model
 
 ### Changed
 
-- simple refactoring - renaming
+- code refactoring - renaming
 - the source codes were moved into the directory `src`
+- the client id and the client secret are now supplied as environmental variables
 
 ### Removed
 
@@ -23,3 +27,4 @@ This changelog documents the changes applied to the previous version of this pro
   - the entire folder `summaries`
   - the entire folder `data`
   - the entire folder `database`
+- all the analysis scripts were removed, currently only data getters are implemented